pldgithub.py: Don't operate on ignored repos

+++ bin/pldgithub.py
@@ -5,6 +5,13 @@ import os
 import json
 import requests
 
+def check_repo(repo):
+    with open(os.path.expanduser("~/ignore"), "r") as f:
+        for line in f:
+            if line.strip() == repo:
+                return 1
+    return 0
+
 if len(sys.argv) < 3 or sys.argv[1] not in ('create', 'delete', 'description'):
     print("""Usage: pldgithub.py create REPO [, REPO2 [, REPO3...]]
    or: pldgithub.py delete REPO [, REPO2 [, REPO3...]]
@@ -21,11 +28,16 @@ if sys.argv[1] == 'create':
             sys.stderr.write("Cannot create {} on github\n".format(newrepo))
 elif sys.argv[1] == 'delete':
     for cannedrepo in [repo.strip() for repo in sys.argv[2:]]:
+        if check_repo(cannedrepo):
+            sys.stderr.write("Ignoring deletion of {} on github\n".format(cannedrepo))
+            continue
         req = requests.delete("https://api.github.com/repos/pld-linux/"+cannedrepo, auth=logpass)
         if not req.status_code == 204:
             sys.stderr.write("Cannot delete {} from github\n".format(cannedrepo))
 else:
     (repo, newdesc) = [arg.strip() for arg in sys.argv[2:4]]
+    if check_repo(repo):
+        raise SystemExit("Ignoring sending summary of {} to github".format(repo))
     req = requests.patch("https://api.github.com/repos/pld-linux/"+repo, auth=logpass,
             data=json.dumps({'name': repo, 'description': newdesc}))
     if not req.status_code == 200:
